Completed PlanVirtualStave; Starting full skeleton BuildThermomechanicalStave
PlanVirtualStave computes the all module fiducial positions in the stave
basis. It exports those values as a 4x3 array of xyz coordinates in the
stave basis. The full virtual stave can then be manipulated using this
array.

Turning towards building the full workflow skeleton for the TM stave,
"BuildThermomechanicalStave," which will be the user's working
interface.