Almost Ready for LS
Mainly Changes:
- PinstoAngle.vi. (calls Findmidpoints-Toolingpinsutility.vi) This Vi
aims to compute the angle stage to stave. Since the new conical tooling
pins, it takes into account two template: one for the left side and one
for the right. Moreover it compute the mid points of each tooling pins
to do the linear fit. The pattern matching procedure has been improved
and results to be much faster.
This VI replaced the old CalculateAngleStagetoStave in Build
Thermomechanical Stave.
-New configuration FIles: StereoSutcliffeStaveCore9-Side1/Side 2 + New
module confi files and new templates